inputs상태와 handleChange 함수를 App.jsx로 올림
