+++ src/components/CardNumbers.jsx
@@ -1,24 +1,6 @@
-import React, { useState } from 'react';
+import React from 'react';
 
-const CardNumbers = () => {
-  const [inputs, setInputs] = useState({
-    cardNumber1: '',
-    cardNumber2: '',
-    cardNumber3: '',
-    cardNumber4: '',
-  });
-
-  function handleChange(event) {
-    const value = event.target.value;
-    if (!isNaN(value)) {
-      // value가 문자열이면 NaN이니까, NaN이 아니면..., 즉 숫자일 때에만!
-      setInputs(old => ({ ...old, [event.target.name]: value }));
-    } else {
-      console.log('문자열이야!', value);
-    }
-  }
-
-  console.log(inputs);
+const CardNumbers = ({ inputs, handleChange }) => {
   return (
     <>
       <label htmlFor="CardNumber1" className="input-title">
